Relocate getIdentifierText helper into shared AST utilities
Moved getIdentifierText from src/plugin/src/printer/optimizations/loop-size-hoisting.js into src/shared/ast-node-helpers.js so printer and AST transform modules can import it without reaching into printer internals.
