+++ src/plugin/src/printer/optimizations/loop-size-hoisting.js
@@ -2,6 +2,8 @@
 // This logic analyses the AST rather than producing Prettier docs, so it lives
 // alongside other printer optimizations instead of the main print pipeline.
 
+import { getIdentifierText } from "../../../../shared/ast-node-helpers.js";
+
 const DEFAULT_SIZE_RETRIEVAL_FUNCTION_SUFFIXES = new Map([
     ["array_length", "len"],
     ["ds_list_size", "size"],
@@ -171,64 +173,9 @@ function buildCachedSizeVariableName(baseName, suffix) {
     return `${baseName}_${normalizedSuffix}`;
 }
 
-function getIdentifierText(node) {
-    if (!node) {
-        return null;
-    }
-
-    if (typeof node === "string") {
-        return node;
-    }
-
-    if (typeof node.name === "string") {
-        return node.name;
-    }
-
-    if (node.type === "Identifier") {
-        return node.name || null;
-    }
-
-    if (node.type === "MemberIndexExpression") {
-        const object = node.object;
-        if (!object || object.type !== "Identifier") {
-            return null;
-        }
-
-        if (!Array.isArray(node.property) || node.property.length !== 1) {
-            return null;
-        }
-
-        const indexNode = node.property[0];
-        const indexText = getIdentifierText(indexNode);
-        if (indexText == null) {
-            return null;
-        }
-
-        return `${object.name}_${indexText}`;
-    }
-
-    if (node.type === "MemberDotExpression") {
-        const object = node.object;
-        const property = node.property;
-
-        if (!object || object.type !== "Identifier" || !property || property.type !== "Identifier") {
-            return null;
-        }
-
-        return `${object.name}_${property.name}`;
-    }
-
-    if (node.type === "Literal" && typeof node.value === "string") {
-        return node.value;
-    }
-
-    return null;
-}
-
 export {
     DEFAULT_SIZE_RETRIEVAL_FUNCTION_SUFFIXES,
     buildCachedSizeVariableName,
     getArrayLengthHoistInfo,
-    getSizeRetrievalFunctionSuffixes,
-    getIdentifierText
+    getSizeRetrievalFunctionSuffixes
 };
