+++ src/plugin/tests/get-identifier-text.test.js
@@ -2,7 +2,7 @@ import assert from "node:assert/strict";
 
 import { describe, it } from "node:test";
 
-import { getIdentifierText } from "../src/printer/optimizations/loop-size-hoisting.js";
+import { getIdentifierText } from "../../shared/ast-node-helpers.js";
 
 describe("getIdentifierText", () => {
     it("returns string arguments unchanged", () => {
